backend/pyproject.toml
[project]
name = "nomitomo-backend"
version = "0.1.0"
description = "Backend for Nomitomo App (FastAPI + Clean Architecture)"
authors = [{name = "Kamontia", email = "kamon@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "pydantic>=2.7.1",
    "pydantic-settings>=2.2.1",
    "alembic>=1.13.1",
    "mangum>=0.17.0",  # For AWS Lambda adapter
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
